26.02 main merge 25.10 main

+++ storage/ndb/rest-server2/server/src/ttl_purge.hpp
@@ -201,7 +201,7 @@ class TTLPurger {
   // minus_sec lets callers encode a past instant (now - minus_sec). The TTL
   // purge scan passes ttl_sec + purge_window to obtain the expiry threshold.
   static Int64 GetNow(unsigned char* encoded_now, bool timestamp,
-                      Uint32 minus_sec = 0);
+                      Uint64 minus_sec = 0);
   bool UpdateLease(const unsigned char* encoded_now);
   bool IsNodeAlive(const unsigned char* encoded_last_active);
   Uint32 AdjustBatchSize(Uint32 curr_batch_size,
